Adding threshold argument for customization (#14)
Adding customization via the threshold argument.

Previously the default threshold for reporting was 90% and was not editable.

This argument --threshold takes a number between 0 and 1 (inclusive) that is used to set the threshold. --threshold all can also be used to return all resources (and set threshold to 0). If unspecified, aws-size will default to 90%.