S3 policy size (#38)
Adds support for checking S3 bucket policy size.

Size limit is 20 KB. AWS does normalization so the 20kb is not necessarily character size of the string representation of the string. aws-size does an approximate normalization on our end from testing multiple bucket policies to come close to the size limit. This will not be an exact normalization and AWS's exact size limit may be slightly different.