Add Ability to save output to a local file (#58)
This PR adds the abliity to save aws-size's run output to a local flag via the --output argument. This can be done by providing a file name after the argument such as --output aws-size-output.json.

README.md has been updated for instructions on how to save output and what metadata is included.